+++ packages/atomic/src/components/facets-v1/atomic-color-facet/atomic-color-facet.tsx
@@ -20,7 +20,7 @@ import {FacetPlaceholder} from '../../facets/atomic-facet-placeholder/atomic-fac
 import {FacetContainer} from '../facet-container/facet-container';
 import {FacetHeader} from '../facet-header/facet-header';
 import {FacetSearchInput} from '../facet-search/facet-search-input';
-import {FacetValueCheckbox} from '../facet-value-checkbox/facet-value-checkbox';
+import {ColorFacetCheckbox} from '../color-facet-checkbox/color-facet-checkbox';
 import {FacetValueBox} from '../facet-value-box/facet-value-box';
 import {FacetShowMoreLess} from '../facet-show-more-less/facet-show-more-less';
 import {FacetSearchMatches} from '../facet-search/facet-search-matches';
@@ -30,6 +30,10 @@ import {
 } from '../facet-search/facet-search-utils';
 import {BaseFacet} from '../facet-common';
 import {FacetValueLabelHighlight} from '../facet-value-label-highlight/facet-value-label-highlight';
+import {
+  getFieldCaptions,
+  getFieldValueCaption,
+} from '../../../utils/field-utils';
 
 /**
  * A facet is a list of values for a certain field occurring in the results, ordered using a configurable criteria (e.g., number of occurrences).
@@ -55,9 +59,8 @@ import {FacetValueLabelHighlight} from '../facet-value-label-highlight/facet-val
  * @part value-label - The facet value label, common for all displays.
  * @part value-count - The facet value count, common for all displays.
  *
- * @part value-checkbox - The facet value checkbox, available when display is 'checkbox'.
- * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
  * @part value-box - The facet value when display is 'box'.
+ * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
  *
  * @part show-more - The show more results button.
  * @part show-less - The show less results button.
@@ -183,6 +186,9 @@ export class AtomicColorFacet
             this.facet.facetSearch.clear();
             return;
           }
+          this.facet.facetSearch.updateCaptions(
+            getFieldCaptions(this.field, this.bindings.i18n)
+          );
           this.facet.facetSearch.updateText(value);
           this.facet.facetSearch.search();
         }}
@@ -192,12 +198,19 @@ export class AtomicColorFacet
   }
 
   private renderValue(facetValue: FacetValue, onClick: () => void) {
-    const displayValue = this.bindings.i18n.t(facetValue.value);
+    const displayValue = getFieldValueCaption(
+      this.facetId!,
+      facetValue.value,
+      this.bindings.i18n
+    );
     const isSelected = facetValue.state === 'selected';
+    const partValue = displayValue
+      .match(new RegExp('-?[_a-zA-Z]+[_a-zA-Z0-9-]*'))
+      ?.toString();
     switch (this.displayValuesAs) {
       case 'checkbox':
         return (
-          <FacetValueCheckbox
+          <ColorFacetCheckbox
             displayValue={displayValue}
             numberOfResults={facetValue.numberOfResults}
             isSelected={isSelected}
@@ -210,7 +223,7 @@ export class AtomicColorFacet
               isSelected={isSelected}
               searchQuery={this.facetState.facetSearch.query}
             ></FacetValueLabelHighlight>
-          </FacetValueCheckbox>
+          </ColorFacetCheckbox>
         );
       case 'box':
         return (
@@ -222,6 +235,10 @@ export class AtomicColorFacet
             onClick={onClick}
             searchQuery={this.facetState.facetSearch.query}
           >
+            <div
+              part={`value-${partValue}`}
+              class="value-box-color w-full h-12 bg-neutral-dark rounded-md mb-2"
+            ></div>
             <FacetValueLabelHighlight
               displayValue={displayValue}
               isSelected={isSelected}
@@ -234,7 +251,7 @@ export class AtomicColorFacet
 
   private renderValuesContainer(children: VNode[]) {
     const classes = `mt-3 ${
-      this.displayValuesAs === 'box' ? 'box-container' : ''
+      this.displayValuesAs === 'box' ? 'box-color-container' : ''
     }`;
     return (
       <ul part="values" class={classes}>