+++ packages/headless/src/controllers/standalone-search-box/headless-standalone-search-box.test.ts
@@ -93,6 +93,7 @@ describe('headless standalone searchBox', () => {
       })),
       redirectTo: state.redirection.redirectTo,
       isLoading: false,
+      isLoadingSuggestions: false,
       analytics: {
         cause: '',
         metadata: null,